Consolidate provider catalog, config and update operations
